bug fixes and polish

+++ RayTracer/MCG_GFX_Lib.h
@@ -12,16 +12,16 @@ namespace MCG
 	/// Initialises window to specified size
 	/// The windowSize parameter specifies how many pixels wide and high the window should be
 	/// \return False if something went wrong
-	bool Init( glm::ivec2 windowSize );
+	bool Init( glm::vec2 windowSize );
 
 	/// Sets every pixel to specified colour
 	/// The colour parameter ranges from 0 to 255. Values outside this range may give unpredictable results.
-	void SetBackground( glm::ivec3 colour );
+	void SetBackground( glm::vec3 colour );
 
 	/// Draws a single pixel to screen
 	/// The position parameter is in pixel-coordinates, ranging from 0,0 to the size of the screen set with the Init function. If coordinates are out of bounds, does nothing.
 	/// The colour parameter ranges from 0 to 255. Values outside this range may give unpredictable results.
-	void DrawPixel( glm::ivec2 position, glm::vec3 colour );
+	void DrawPixel( glm::vec2 position, glm::vec3 colour );
 
 	/// Displays graphics to screen and keeps window open until user requests exit (pressing escape key or closing window)
 	int ShowAndHold();
